+++ application/spring-boot-app/src/main/java/org/mobilitydata/gtfsvalidator/Main.java
@@ -16,24 +16,12 @@
 
 package org.mobilitydata.gtfsvalidator;
 
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import java.io.IOException;
-
-@Controller
-public class ApiController {
-    private final ServiceManager serviceManager;
-
-    public ApiController(ServiceManager serviceManager) {
-        this.serviceManager = serviceManager;
-    }
-
-    @RequestMapping(value = "/report", method = RequestMethod.GET)
-    @ResponseBody
-    public String getReport() throws IOException {
-        return serviceManager.getReport();
+@SpringBootApplication
+public class Main {
+    public static void main(String[] args) {
+        SpringApplication.run(Main.class, args);
     }
 }
